UPDATE: Update menus and drinks page
This commit updates the restaurant menus pdf files in the menus folder to reflect the 2023 and 2024 menu options. It also updates the drinks.html page to change "restaurant" to "Italian restaurant" in the meta tags and text to reflect that the establishment serves Italian cuisine. Minor styling tweaks were also made to index.html.